pyproject.toml
[project]
name = "actuator-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.14"
dependencies = [
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "ipykernel>=7.2.0",
    "numpy>=2.4.4",
    "ollama>=0.6.1",
    "openai>=2.31.0",
    "openai-agents>=0.13.6",
    "psycopg>=3.3.3",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "sqlmodel>=0.0.38",
    "uvicorn>=0.44.0",
]
